requirements-local.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
djangorestframework-simplejwt==5.3.0
django-filter==23.3
Pillow==10.0.1
flet==0.21.2
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.6.0